Renamed BookmarkedCommentResource to CommentCardResource
